refactor(runtime,network,task,agent): ExitReason/WorkerRole 枚举下沉 common/runtime + 线上字段定型
审计批次 C 第二程 P2-3 / P2-4：

- WorkerRole 自 task/cpp/worker_manager.h 下沉 common/runtime/cpp/worker_role.h；
  ExitReason 自 agent/cpp/worker_agent.h 下沉 common/runtime/cpp/exit_reason.h
  （与 error_types.h 同层）。两枚举均为线上消息字段值（RegisterMessage.role_ /
  WorkerExitMessage.exit_reason_），network 不允许反向依赖 task/agent——
  下沉后 network 与 task/agent 共用单一定义，agent 层 include 收敛
  （worker_agent.h 不再为取枚举 include worker_manager.h）
- RegisterMessage.role_ / WorkerExitMessage.exit_reason_ / WorkerAgent.role_
  自裸 uint8_t 定型为枚举（uint8 同宽，bitsery value1b wire 编码不变）；
  worker_agent 生产点与 master_agent 注册路径的反向 static_cast 消除
- master 注册路径对越界 role 值的显式校验保留（越界 wire 值定型后经
  static_cast 注入，拒绝行为由既有测试 master_agent_test 越界负例继续锁定）

字节级测试锚定（message_protocol_test IdentityEnumWireTest）：
- RegisterMessage 全空变长域下 role_=STORAGE_ONLY 为 payload 末字节 0x01 直通
- WorkerExitMessage exit_reason_=MASTER_LOST 为末字节 0x02 直通
- 两消息 round-trip 枚举值还原

task 模块 BUILD 显式补 common/runtime 依赖（头文件下沉引入）。
全量 C++ 单测 113/113 通过。

+++ src/agent/cpp/worker_agent.cpp
@@ -50,13 +50,13 @@ WorkerAgent::WorkerAgent(uint64_t worker_id, const CMString& master_host, uint16
     // role 静态身份（注册时设定，不可变更）：仅 hybrid / storage_only；
     // 非法值 WARN 回退 hybrid。
     if (role == "storage_only") {
-        role_ = static_cast<uint8_t>(WorkerRole::STORAGE_ONLY);
+        role_ = WorkerRole::STORAGE_ONLY;
     } else {
         if (!role.empty() && role != "hybrid") {
             WARN("Unknown worker role '{}' (expected hybrid|storage_only), "
                  "falling back to hybrid", role);
         }
-        role_ = static_cast<uint8_t>(WorkerRole::HYBRID);
+        role_ = WorkerRole::HYBRID;
     }
 }
 
@@ -597,7 +597,7 @@ void WorkerAgent::send_register_message() {
     auto dsp = data_server_port_;
     auto attr_count = attributes_.size();
     INFO("RegisterMessage sent with data_server_port={}, attributes={}, role={}",
-         dsp, attr_count, role_ == static_cast<uint8_t>(WorkerRole::STORAGE_ONLY)
+         dsp, attr_count, role_ == WorkerRole::STORAGE_ONLY
                                ? "storage_only" : "hybrid");
 }
 
@@ -1332,7 +1332,7 @@ void WorkerAgent::initiate_shutdown(ExitReason reason, const CMString& detail) {
         if (conn != 0 && reactor_) {
             WorkerExitMessage msg;
             msg.worker_id_ = worker_id_;
-            msg.exit_reason_ = static_cast<uint8_t>(reason);
+            msg.exit_reason_ = reason;
             reactor_->send(conn, msg);
         }
     }
